pyproject.toml
[project]
name = "tpm2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openreview-py>=1.50.0",
    "pyyaml>=6.0.2",
]
